project-api/requirements.txt
fastapi==0.89.1
uvicorn[standard]==0.20.0
pydantic==1.10.4
email-validator==1.3.1
SQLAlchemy==1.4.46
psycopg2-binary==2.9.5
alembic==1.9.2
tenacity==8.2.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
starlette==0.22.0
postmarker==1.0
requests==2.28.2
Jinja2==3.1.2
sentry-sdk==1.30.0